% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2mu.R
\name{alpha2mu.base}
\alias{alpha2mu.base}
\title{Compute expected log10 parasite densities, \code{mu}, as a function of the age of infection \code{alpha}}
\usage{
\method{alpha2mu}{base}(alpha, W, par)
}
\arguments{
\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{par}{a \link{list} that defines a model}
}
\value{
mean log10 parasite densities
}
\description{
Compute mean log10 parasites for a model with no immunity
}
