% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_complex.R
\name{Bda}
\alias{Bda}
\title{Compute the CDF and PDF of parasite densities in a host cohort}
\usage{
Bda(
  meshX,
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu_base(),
  pRBC = par_lRBC_static(),
  pSig = par_sigma_abc(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{meshX}{a mesh over parasite densities}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length meshX
}
\description{
Compute the CDF and PDF of parasite densities in a host cohort
}
