% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_parasites_complex.R
\name{Bdensity}
\alias{Bdensity}
\title{Compute \eqn{B_\tau(a | h)}}
\usage{
Bdensity(
  meshX,
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  mu_par = par_mu2dens_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{meshX}{a mesh over parasite densities}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{alpha2mu}}

\item{mu_par}{parameters to compute \link{d_mu2density}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length meshX
}
\description{
This function first computes the PDF for simple infections, \link{d_Pdensity}, and
then computes the CDF for the full distribution, summed from MoI = 1 up to a large number
using convolutions.
}
