% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-nb.R
\name{binnedCounts.nb}
\alias{binnedCounts.nb}
\title{Detection of infection given parasitemia}
\usage{
\method{binnedCounts}{nb}(xi, a, bins, Cpar = par_nbCounts())
}
\arguments{
\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{bins}{break points for binning counts data}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
