% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{binnedCounts.pois}
\alias{binnedCounts.pois}
\title{Detection of infection given parasitemia}
\usage{
\method{binnedCounts}{pois}(xi, a, bins, Cpar = par_poisCounts())
}
\arguments{
\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{bins}{breakpoints for binning data}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
