% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-interface.R
\name{binnedCounts}
\alias{binnedCounts}
\title{Detection of infection given parasitemia}
\usage{
binnedCounts(xi, a, bins, Cpar = par_nbCounts())
}
\arguments{
\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{bins}{breakpoints for summarizing outputs}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
