% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-interface.R
\name{d_mu2density}
\alias{d_mu2density}
\title{The density function for parasite densities in a simple malaria infection}
\usage{
d_mu2density(xi, mu, par_mu2dens = par_mu2dens_beta(), bvm = 13)
}
\arguments{
\item{xi}{vector of quantiles for \eqn{log_{10}} parasite densities}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{par_mu2dens}{parameters to compute parasite densities as a function of mu}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The density function for parasite densities in a simple malaria infection
}
