% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{dAoI}
\alias{dAoI}
\title{The density function for the age of infection (AoI)}
\usage{
dAoI(alpha, a, FoIpar, hhat = NULL, tau = 0, r = 1/200)
}
\arguments{
\item{alpha}{the parasite age of infection}

\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The density function for the age of infection (AoI)
}
