% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoy.R
\name{dAoYda}
\alias{dAoYda}
\title{Compute the derivatives of the approximate moments of the AoY dynamically}
\usage{
dAoYda(a, M, par, FoIpar)
}
\arguments{
\item{a}{the host age}

\item{M}{the state variables}

\item{par}{the parameters}

\item{FoIpar}{parameters that define an FoI function}
}
\value{
the derivatives of the MoI and the AoI as a \link{list}
}
\description{
The
}
