% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoy.R
\name{dAoYN}
\alias{dAoYN}
\title{The density function for the youngest age of N infections}
\usage{
dAoYN(N, a, FoIpar, hhat = NULL, tau = 0, r = 1/200)
}
\arguments{
\item{N}{the number of infections}

\item{a}{the age of a cohort}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate of a simple infection}
}
\value{
a \link{numeric} vector of length a + 1
}
\description{
The density function for the youngest age of N infections
}
