% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{dbeta1}
\alias{dbeta1}
\title{Density function for the beta distribution, an alternative parameterization}
\usage{
dbeta1(x, mu, pSig = par_sigma.0())
}
\arguments{
\item{x}{a vector of quantiles}

\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{pSig}{parameters to dispatch the S3 function \link{sigma}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The beta distribution is parameterized using the mean and a function \code{sigma} that computes the variance
as a function of the mean
}
