% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-pois.R
\name{dCounts.pois}
\alias{dCounts.pois}
\title{Detection of infection given parasitemia}
\usage{
\method{dCounts}{pois}(hatxi, xi, a = 0, Cpar = par_poisCounts())
}
\arguments{
\item{hatxi}{mean log10 parasite densities}

\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
