% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_counts.R
\name{dCountsBa}
\alias{dCountsBa}
\title{Compute the density of parasite counts in complex infections}
\usage{
dCountsBa(
  a,
  FoIpar,
  bins = 1,
  dx = 0.1,
  hhat = 1,
  tau = 0,
  r = 1/200,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  mu_par = par_mu2dens_beta(),
  pWda = par_Wda_none(),
  Cpar = par_nbCounts()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{bins}{a set of break points for computing counts}

\item{dx}{the width of the mesh for computing the CDF of parasite densities}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{alpha2mu}}

\item{mu_par}{parameters defining parasite densities as a function of the mu}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{Cpar}{parameters that define a detection function}
}
\value{
a \link{list}
}
\description{
Compute the density of parasite counts in complex infections
}
