% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{dDensityPa}
\alias{dDensityPa}
\title{The density function for parasite densities from a host cohort population of age a}
\usage{
dDensityPa(
  x,
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu_base(),
  pRBC = par_lRBC_static(),
  pSig = par_sigma_abc(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{x}{host cohort age}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma_mu}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The density function for parasite densities from a host cohort population of age a
}
