% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{dDensityPaConvolve2}
\alias{dDensityPaConvolve2}
\title{The density function for the sum of two infections}
\usage{
dDensityPaConvolve2(
  x,
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0(),
  pWda = par_Wda.delta()
)
}
\arguments{
\item{x}{host cohort age}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} value
}
\description{
The density function for the sum of two infections
}
