% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{dDensityPalpha}
\alias{dDensityPalpha}
\title{The density function for parasite densities in a simple malaria infection of age alpha}
\usage{
dDensityPalpha(
  x,
  alpha,
  W = 0,
  a = 0,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0()
)
}
\arguments{
\item{x}{a vector of quantiles}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The density function for parasite densities in a simple malaria infection of age alpha
}
