% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-nb.R
\name{Detect.nb}
\alias{Detect.nb}
\title{Detection of infection given parasitemia}
\usage{
\method{Detect}{nb}(xi, a = 0, Cpar = par_nbCounts())
}
\arguments{
\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
