% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-interface.R
\name{Detect}
\alias{Detect}
\title{Detection of infection given parasitemia}
\usage{
Detect(xi, a = 0, Cpar = par_nbCounts())
}
\arguments{
\item{xi}{mean \eqn{\log_{10}} parasite densities}

\item{a}{host cohort age}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
