% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_detect.R
\name{DetectBda}
\alias{DetectBda}
\title{Detection of infection given parasitemia}
\usage{
DetectBda(
  a,
  FoIpar,
  dx = 0.1,
  tau = 0,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  par_mu2dens = par_mu2dens_beta(),
  pWda = par_Wda_none(),
  pC = par_nbCounts()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{dx}{width of the mesh}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{alpha2mu}}

\item{par_mu2dens}{parameters to compute \link{d_mu2density}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{pC}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
