% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DetectPM}
\alias{DetectPM}
\title{Detection of infection given parasitemia}
\usage{
DetectPM(
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0(),
  pWda = par_Wda.delta(),
  pC = par_nbCounts()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{pC}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
