% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{dMoIda}
\alias{dMoIda}
\title{Compute the first derivatives for the queuing model M/M/infinity}
\usage{
dMoIda(a, M, p, FoIpar)
}
\arguments{
\item{a}{the host age}

\item{M}{the state variables}

\item{p}{the parameters}

\item{FoIpar}{parameters that define an FoI function}
}
\value{
the derivatives as a \link{list}
}
\description{
Compute the first derivatives for the queuing model M/M/infinity
}
