% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{FoI}
\alias{FoI}
\title{Trace Function - Force of Infection}
\usage{
FoI(a, par, tau = 0, hhat = NULL)
}
\arguments{
\item{a}{host age}

\item{par}{a formatted list}

\item{tau}{cohort birthday}

\item{hhat}{overrides the value of hbar in par}
}
\value{
\link{numeric}
}
\description{
Trace Function - Force of Infection
}
