% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{make_FoI}
\alias{make_FoI}
\title{Trace Function - Force of Infection}
\usage{
make_FoI(par, maxAge = 50)
}
\arguments{
\item{par}{a formatted list}

\item{maxAge}{age in years to normalize over}
}
\value{
\link{numeric}
}
\description{
Trace Function - Force of Infection
}
