% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{meanMoI}
\alias{meanMoI}
\title{Compute the mean MoI directly}
\usage{
meanMoI(a, FoIpar, hhat = NULL, tau = 0, r = 1/200)
}
\arguments{
\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Compute the mean MoI directly
}
