% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{MoIDistPlot}
\alias{MoIDistPlot}
\title{Plot the distribution of the MoI at time t}
\usage{
MoIDistPlot(moi, t, clr1 = "red")
}
\arguments{
\item{moi}{the mean moi}

\item{t}{the time}

\item{clr1}{the color}
}
\description{
Plot the distribution of the MoI at time t
}
