% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_parasites_clones.R
\name{moment_Pdensity}
\alias{moment_Pdensity}
\title{Compute the moments of P_density}
\usage{
moment_Pdensity(
  a,
  FoIpar,
  tau = 0,
  n = 1,
  dt = 0.1,
  hhat = 1,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_mu2dens = par_mu2dens_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{n}{the moment to compute}

\item{dt}{the mesh size over xi}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{alpha2mu}}

\item{par_mu2dens}{parameters to compute \link{d_mu2density}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
Compute \deqn{P_\tau(a| h) \sim f_P(\xi; a, \tau |h ) = \int_0^a \Omega(\xi|F_\mu(\alpha)) \; f_A(\alpha; a, \tau | h) d\alpha}
}
