% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{momentAoI}
\alias{momentAoI}
\title{Compute the moments for the AoI density function for a cohort of age a}
\usage{
momentAoI(a, FoIpar, hhat = 5/365, tau = 0, r = 1/200, n = 1)
}
\arguments{
\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{n}{the moment desired}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Compute the moments for the AoI density function for a cohort of age a
}
