% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_parasites_complex.R
\name{p_Bdensity}
\alias{p_Bdensity}
\title{Call \link{Bdensity} and return the density vector}
\usage{
p_Bdensity(
  meshX,
  a,
  FoIpar,
  tau = 0,
  hhat = 1,
  r = 1/200,
  RBC_par = par_lRBC_static(),
  Fmu_par = par_Fmu_base(),
  mu_par = par_mu2dens_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{meshX}{a mesh over parasite densities}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{RBC_par}{parameters to compute \link{log10RBC}}

\item{Fmu_par}{parameters to compute \link{alpha2mu}}

\item{mu_par}{parameters to compute \link{d_mu2density}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a numeric vector of length meshX
}
\description{
Call \link{Bdensity} and return the density vector
}
