% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-beta.R
\name{p_mu2density.beta}
\alias{p_mu2density.beta}
\title{Modified beta distribution, distribution function}
\usage{
\method{p_mu2density}{beta}(xi, mu, par_mu2dens = par_mu2dens_beta(), bvm = 13)
}
\arguments{
\item{xi}{a vector of probabilities for \eqn{log_{10}} parasite densities}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{par_mu2dens}{parameters defining parasite densities as a function of the mu}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
Modified beta distribution, distribution function
}
