% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoy.R
\name{pAoY}
\alias{pAoY}
\title{The distribution function for the age of the youngest infection (AoY)}
\usage{
pAoY(alpha, a, FoIpar, hhat = NULL, tau = 0, r = 1/200)
}
\arguments{
\item{alpha}{the age of infection}

\item{a}{the age of a cohort}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate of a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The distribution function for the age of the youngest infection (AoY)
}
