% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2mu-chronic.R
\name{par_Fmu_chronic}
\alias{par_Fmu_chronic}
\title{Set up parameters for \link{alpha2mu.chronic}}
\usage{
par_Fmu_chronic(tildeb = 10.3, tildel = 2, Sa = 0.0033)
}
\arguments{
\item{tildeb}{The maximum expected log10 parasite densities}

\item{tildel}{The minimum expected log10 parasite densities}

\item{Sa}{The decline in mu with respect to alpha}
}
\value{
a \link{list}
}
\description{
Set up parameters for \link{alpha2mu.chronic}
}
