% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbc.R
\name{par_lRBC.0}
\alias{par_lRBC.0}
\title{Set up parameters for \link{log10RBC.0}}
\usage{
par_lRBC.0(lRBCmax = 13)
}
\arguments{
\item{lRBCmax}{the maximum}
}
\value{
a \link{list}
}
\description{
Set up parameters for \link{log10RBC.0}
}
