% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-beta.R
\name{par_mu2dens_beta}
\alias{par_mu2dens_beta}
\title{The quantile function for parasite densities in a simple malaria infection}
\usage{
par_mu2dens_beta(par_sigma = par_sigma_abc())
}
\arguments{
\item{par_sigma}{parameters to compute sigma_mu}
}
\value{
a compound \link{list}
}
\description{
The quantile function for parasite densities in a simple malaria infection
}
