% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{par_nbCounts}
\alias{par_nbCounts}
\title{Detection of infection given parasitemia}
\usage{
par_nbCounts(q = 6, sz = 0.31, bvm = par_lRBC_static())
}
\arguments{
\item{q}{the fraction of blood examined}

\item{sz}{the negative binomial size parameter}

\item{bvm}{blood volume as log10 red blood cells}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
