% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-pois.R
\name{par_poisCounts}
\alias{par_poisCounts}
\title{Detection of infection given parasitemia}
\usage{
par_poisCounts(q = 6, bvm = par_lRBC_static())
}
\arguments{
\item{q}{the fraction of blood volume sampled}

\item{bvm}{blood volume as log10 red blood cells}
}
\value{
par a \link{list}
}
\description{
Detection of infection given parasitemia
}
