% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{par_sigma.0}
\alias{par_sigma.0}
\title{Parameters to configure \link{sigma.0}}
\usage{
par_sigma.0(aa = 3.11, bb = 2.14, cc = 1.74)
}
\arguments{
\item{aa}{a shape parameter}

\item{bb}{a shape parameter}

\item{cc}{a shape parameter}
}
\value{
a \link{list}
}
\description{
Parameters to configure \link{sigma.0}
}
