% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_tracking.R
\name{par_Wda.delta}
\alias{par_Wda.delta}
\title{Make a parameter set for \link{Wda.delta}}
\usage{
par_Wda.delta(delta = 0.001)
}
\arguments{
\item{delta}{a decay rate}
}
\value{
a \link{list}
}
\description{
Make a parameter set for \link{Wda.delta}
}
