% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{pCounts}
\alias{pCounts}
\title{cumulative density function of counts}
\usage{
pCounts(hatxi, xi, a = 0, Cpar = par_nbCounts())
}
\arguments{
\item{hatxi}{mean log10 parasite densities}

\item{xi}{mean log10 parasite densities}

\item{a}{host cohort age}

\item{Cpar}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
cumulative density function of counts
}
