% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{pCountsPa}
\alias{pCountsPa}
\title{Compute the distribution of parasite counts for simple infections}
\usage{
pCountsPa(
  a,
  FoIpar,
  bins = NULL,
  dx = 0.1,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0(),
  pWda = par_Wda.delta(),
  Cpar = par_nbCounts()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{bins}{a set of break points for computing counts}

\item{dx}{the width of the mesh for computing the CDF of parasite densities}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{Cpar}{parameters that define a detection function}
}
\value{
a \link{list}
}
\description{
Compute the distribution of parasite counts for simple infections
}
