% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{pDensityPmu}
\alias{pDensityPmu}
\title{The density function for parasite densities as a function of the mean}
\usage{
pDensityPmu(xi, mu, a = 0, pRBC = par_lRBC_static(), pSig = par_sigma_abc())
}
\arguments{
\item{xi}{a vector of probabilities}

\item{mu}{the expected value for log10 parasite densities}

\item{a}{host cohort age}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma_mu}}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The density function for parasite densities as a function of the mean
}
