% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2density.R
\name{q_alpha2density}
\alias{q_alpha2density}
\title{The quantile function for parasite densities in a simple malaria infection of age alpha}
\usage{
q_alpha2density(
  x,
  alpha,
  W = 0,
  a = 0,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_mu2dens = par_mu2dens_beta()
)
}
\arguments{
\item{x}{a vector of quantiles}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{alpha2mu}}

\item{par_mu2dens}{parameters to compute \link{q_mu2density}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The quantile function for parasite densities in a simple malaria infection of age alpha
}
