% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-interface.R
\name{q_mu2density}
\alias{q_mu2density}
\title{The quantile function for parasite densities in a simple malaria infection}
\usage{
q_mu2density(xi, mu, par_mu2dens = par_mu2dens_beta(), bvm = 13)
}
\arguments{
\item{xi}{a vector of quantiles}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{par_mu2dens}{parameters defining parasite densities as a function of the mu}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The quantile function for parasite densities in a simple malaria infection
}
