% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{qbeta1}
\alias{qbeta1}
\title{The quantile function for the beta distribution, an alternative parameterization}
\usage{
qbeta1(x, mu, pSig = par_sigma.0())
}
\arguments{
\item{x}{a vector of quantiles}

\item{mu}{the mean value for the distribution (0 <= mu <= 1)}

\item{pSig}{parameters to dispatch the S3 function \link{sigma}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The quantile function for the beta distribution, an alternative parameterization
}
