% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{qDensityPalpha}
\alias{qDensityPalpha}
\title{The quantile function for parasite densities in a simple malaria infection of age alpha}
\usage{
qDensityPalpha(
  x,
  alpha,
  W = 0,
  a = 0,
  pMu = par_alpha2mu_base(),
  pRBC = par_lRBC_static(),
  pSig = par_sigma_abc()
)
}
\arguments{
\item{x}{a vector of quantiles}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma_mu}}
}
\value{
a \link{numeric} vector of length(x)
}
\description{
The quantile function for parasite densities in a simple malaria infection of age alpha
}
