% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-beta.R
\name{r_mu2density.beta}
\alias{r_mu2density.beta}
\title{Modified beta distribution, random numbers}
\usage{
\method{r_mu2density}{beta}(n, mu, par_mu2dens = par_mu2dens_beta(), bvm = 13)
}
\arguments{
\item{n}{the number of observations}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{par_mu2dens}{parameters defining parasite densities as a function of the mu}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The beta distribution, parameterized by the mean and variance, modified to return
a number between 0 and bvm (the number of red blood cells)
}
