% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2density-interface.R
\name{r_mu2density}
\alias{r_mu2density}
\title{Random generation of parasite densities from a simple malaria infection}
\usage{
r_mu2density(n, mu, par_mu2dens = par_mu2dens_beta(), bvm = 13)
}
\arguments{
\item{n}{the number of observations}

\item{mu}{the expected value for \eqn{log_{10}} parasite densities}

\item{par_mu2dens}{parameters defining parasite densities as a function of the mu}

\item{bvm}{blood volume as \eqn{log_{10}} red blood cells}
}
\value{
a \link{numeric} vector of length(n)
}
\description{
Random generation of parasite densities from a simple malaria infection
}
