% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{rAoI}
\alias{rAoI}
\title{The random generation function for the age of infection (AoI)}
\usage{
rAoI(N, a, FoIpar, hhat = NULL, tau = 0, r = 1/200, alphamin = 0)
}
\arguments{
\item{N}{the number of observations}

\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{alphamin}{the minimum value of the AoI to return}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
The random generation function for the age of infection (AoI)
}
