% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{rDensityPalpha}
\alias{rDensityPalpha}
\title{The distribution function for parasite densities in a simple malaria infection of age alpha}
\usage{
rDensityPalpha(
  n,
  alpha,
  W = 0,
  a = 0,
  pMu = par_alpha2mu_base(),
  pRBC = par_lRBC_static(),
  pSig = par_sigma_abc()
)
}
\arguments{
\item{n}{the number of observations}

\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{a}{host cohort age}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to compute \link{sigma}}
}
\value{
a \link{numeric} vector of length(n)
}
\description{
The distribution function for parasite densities in a simple malaria infection of age alpha
}
