% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{rDensityPmu}
\alias{rDensityPmu}
\title{Random generation of parasite densities from a simple malaria infection}
\usage{
rDensityPmu(n, mu, a = 0, pRBC = par_lRBC.0(), pSig = par_sigma.0())
}
\arguments{
\item{n}{the number of observations}

\item{mu}{the expected value for log10 parasite densities}

\item{a}{host cohort age}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}
}
\value{
a \link{numeric} vector of length(n)
}
\description{
Random generation of parasite densities from a simple malaria infection
}
