% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2mu.R
\name{salpha2mu}
\alias{salpha2mu}
\title{Compute mean, expected parasite densities \code{mu} for multiple values of the age of infection \code{alpha}}
\usage{
salpha2mu(alpha, W, par)
}
\arguments{
\item{alpha}{a parasite's age of infection}

\item{W}{immune tracking variables}

\item{par}{a \link{list} that defines a model}
}
\value{
mean log10 parasite densities, a \link{numeric} vector of length(alpha)
}
\description{
Compute mean, expected parasite densities \code{mu} for multiple values of the age of infection \code{alpha}
}
