% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moi.R
\name{solveMMinfty}
\alias{solveMMinfty}
\title{Solve the queuing model M/M/infinity}
\usage{
solveMMinfty(h, FoIpar, r = 1/200, tau = 0, Tmax = 730, dt = 1)
}
\arguments{
\item{h}{the force of infection}

\item{FoIpar}{a FoI trace function}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{Tmax}{The maximum runtime (in days)}

\item{dt}{The output frequency (in days)}
}
\value{
a \link{list} with the orbits by name
}
\description{
Solve the queuing model M/M/infinity
}
