% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_tracking.R
\name{Wda.delta}
\alias{Wda.delta}
\title{Compute immune tracking variables as a function of host age and exposure}
\usage{
\method{Wda}{delta}(a, FoIpar, hhat = NULL, tau = 0, par = par_Wda.0())
}
\arguments{
\item{a}{age of a host cohort}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{overrides the value of hbar in par}

\item{tau}{cohort birthday}

\item{par}{parameters in a \link{list}}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Compute immune tracking variables as a function of host age and exposure
}
