% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{zda}
\alias{zda}
\title{Compute the density of parasite infections in a host cohort}
\usage{
zda(alpha, a, FoIpar, hhat = NULL, tau = 0, r = 1/200)
}
\arguments{
\item{alpha}{a parasites age of infection}

\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(alpha)
}
\description{
Solutions to the time and age version of the Ross-Macdonald model.
}
