% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2mu.R
\name{alpha2mu}
\alias{alpha2mu}
\title{Compute mean, expected parasite densities \code{mu} as a function of the age of infection \code{alpha}}
\usage{
alpha2mu(alpha, W, par)
}
\arguments{
\item{alpha}{the age of a parasite infection}

\item{W}{the immune tracking variables}

\item{par}{a \link{list} that defines a model}
}
\value{
mean log10 parasite densities, a \link{numeric} vector of length(alpha)
}
\description{
Compute mean, expected parasite densities \code{mu} as a function of the age of infection \code{alpha}
}
