% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{dDensityPaConvolve2}
\alias{dDensityPaConvolve2}
\title{The density function for the sum of two infections}
\usage{
dDensityPaConvolve2(
  x,
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  par_RBC = par_lRBC_static(),
  par_Fmu = par_Fmu_base(),
  par_mu2dens = par_mu2dens_beta(),
  pWda = par_Wda_none()
)
}
\arguments{
\item{x}{host cohort age}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{par_RBC}{parameters to compute \link{log10RBC}}

\item{par_Fmu}{parameters to compute \link{alpha2mu}}

\item{par_mu2dens}{parameters to compute \link{d_mu2density}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} value
}
\description{
The density function for the sum of two infections
}
