% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DetectBda}
\alias{DetectBda}
\title{Detection of infection given parasitemia}
\usage{
DetectBda(
  a,
  FoIpar,
  dx = 0.1,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  pMu = par_alpha2mu_base(),
  pRBC = par_lRBC_static(),
  pSig = par_sigma_abc(),
  pWda = par_Wda_none(),
  pC = par_nbCounts()
)
}
\arguments{
\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{dx}{width of the mesh}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma_mu}}

\item{pWda}{parameters to dispatch \link{Wda}}

\item{pC}{parameters that define a detection function}
}
\value{
binary detection result
}
\description{
Detection of infection given parasitemia
}
