% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_simple.R
\name{qDensityPmu}
\alias{qDensityPmu}
\title{The quantile function for parasite densities in a simple malaria infection}
\usage{
qDensityPmu(xi, mu, a = 0, pRBC = par_lRBC.0(), pSig = par_sigma.0())
}
\arguments{
\item{xi}{a vector of quantiles}

\item{mu}{the expected value for log10 parasite densities}

\item{a}{host cohort age}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}
}
\value{
a \link{numeric} vector of length(xi)
}
\description{
The quantile function for parasite densities in a simple malaria infection
}
