% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_complex.R
\name{rBda}
\alias{rBda}
\title{Random generation for parasite densities in a host cohort}
\usage{
rBda(
  N,
  a,
  FoIpar,
  hhat = NULL,
  r = 1/200,
  tau = 0,
  alphamin = 7,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0(),
  pWda = par_Wda.delta()
)
}
\arguments{
\item{N}{number of observations}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}

\item{tau}{the cohort birthday}

\item{alphamin}{the minimum value of alpha allowed}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a \link{numeric} vector of length N
}
\description{
Random generation for parasite densities in a host cohort
}
