% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasite_densities_complex.R
\name{rRda}
\alias{rRda}
\title{Random generation for M parasite densities in a host cohort with MoI}
\usage{
rRda(
  M,
  R,
  a,
  FoIpar,
  hhat = NULL,
  tau = 0,
  r = 1/200,
  alphamin = 7,
  pMu = par_alpha2mu.0(),
  pRBC = par_lRBC.0(),
  pSig = par_sigma.0(),
  pWda = par_Wda.delta()
)
}
\arguments{
\item{M}{the MoI}

\item{R}{the number of observations}

\item{a}{host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{hhat}{a local scaling parameter for the FoI}

\item{tau}{the cohort birthday}

\item{r}{the clearance rate for a simple infection}

\item{alphamin}{the minimum value of alpha allowed}

\item{pMu}{parameters to compute \link{alpha2mu}}

\item{pRBC}{parameters to compute \link{log10RBC}}

\item{pSig}{parameters to dispatch \link{sigma}}

\item{pWda}{parameters to dispatch \link{Wda}}
}
\value{
a R by M \link{matrix}
}
\description{
Random generation for M parasite densities in a host cohort with MoI
}
