% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{truePRa}
\alias{truePRa}
\title{Compute the true PR in a cohort as a function of age and exposure}
\usage{
truePRa(a, FoIpar, tau = 0, hhat = 1, r = 1/200)
}
\arguments{
\item{a}{the host cohort age}

\item{FoIpar}{parameters that define an FoI function}

\item{tau}{the cohort birthday}

\item{hhat}{a local scaling parameter for the FoI}

\item{r}{the clearance rate for a simple infection}
}
\value{
a \link{numeric} vector of length(a)
}
\description{
Compute the true PR in a cohort as a function of age and exposure
}
